backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
pydantic==2.4.2
pydantic-settings==2.0.3
